iox-#1345 Rename platform_building_blocks to building_blocks
