+++ iceoryx_hoofs/CMakeLists.txt
@@ -58,8 +58,8 @@ iox_add_library(
         source/cxx/unique_id.cpp
         source/error_handling/error_handler.cpp
         source/error_handling/error_handling.cpp
-        source/log/platform_building_blocks/console_logger.cpp
-        source/log/platform_building_blocks/logger.cpp
+        source/log/building_blocks/console_logger.cpp
+        source/log/building_blocks/logger.cpp
         source/posix_wrapper/access_control.cpp
         source/posix_wrapper/file_lock.cpp
         source/posix_wrapper/mutex.cpp
