+++ iceoryx_hoofs/include/iceoryx_hoofs/internal/log/building_blocks/logformat.inl
@@ -17,7 +17,7 @@
 #ifndef IOX_HOOFS_PLATFORM_BUILDING_BLOCKS_LOGFORMAT_INL
 #define IOX_HOOFS_PLATFORM_BUILDING_BLOCKS_LOGFORMAT_INL
 
-#include "iceoryx_hoofs/log/platform_building_blocks/logformat.hpp"
+#include "iceoryx_hoofs/log/building_blocks/logformat.hpp"
 
 #include <atomic>
 #include <cstdint>
