+++ iceoryx_hoofs/include/iceoryx_hoofs/internal/log/building_blocks/logger.inl
@@ -18,7 +18,7 @@
 #ifndef IOX_HOOFS_PLATFORM_BUILDING_BLOCKS_LOGGER_INL
 #define IOX_HOOFS_PLATFORM_BUILDING_BLOCKS_LOGGER_INL
 
-#include "iceoryx_hoofs/log/platform_building_blocks/logger.hpp"
+#include "iceoryx_hoofs/log/building_blocks/logger.hpp"
 
 #include <atomic>
 #include <cstdint>
