+++ iceoryx_hoofs/include/iceoryx_hoofs/log/building_blocks/console_logger.hpp
@@ -19,7 +19,7 @@
 #define IOX_HOOFS_PLATFORM_BUILDING_BLOCKS_CONSOLE_LOGGER_HPP
 
 #include "iceoryx_hoofs/iceoryx_hoofs_types.hpp"
-#include "iceoryx_hoofs/log/platform_building_blocks/logformat.hpp"
+#include "iceoryx_hoofs/log/building_blocks/logformat.hpp"
 
 #include <atomic>
 #include <cstdint>
@@ -125,6 +125,6 @@ class ConsoleLogger
 } // namespace pbb
 } // namespace iox
 
-#include "iceoryx_hoofs/internal/log/platform_building_blocks/console_logger.inl"
+#include "iceoryx_hoofs/internal/log/building_blocks/console_logger.inl"
 
 #endif // IOX_HOOFS_PLATFORM_BUILDING_BLOCKS_CONSOLE_LOGGER_HPP
