+++ iceoryx_hoofs/include/iceoryx_hoofs/log/building_blocks/logger.hpp
@@ -107,6 +107,6 @@ class Logger : public BaseLogger
 } // namespace pbb
 } // namespace iox
 
-#include "iceoryx_hoofs/internal/log/platform_building_blocks/logger.inl"
+#include "iceoryx_hoofs/internal/log/building_blocks/logger.inl"
 
 #endif // IOX_HOOFS_PLATFORM_BUILDING_BLOCKS_LOGGER_HPP
