+++ iceoryx_hoofs/source/log/building_blocks/console_logger.cpp
@@ -15,7 +15,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "iceoryx_hoofs/log/platform_building_blocks/console_logger.hpp"
+#include "iceoryx_hoofs/log/building_blocks/console_logger.hpp"
 #include "iceoryx_hoofs/cxx/attributes.hpp"
 #include "iceoryx_platform/time.hpp"
