+++ iceoryx_hoofs/source/log/building_blocks/logger.cpp
@@ -16,7 +16,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "iceoryx_hoofs/log/platform_building_blocks/logger.hpp"
+#include "iceoryx_hoofs/log/building_blocks/logger.hpp"
 #include "iceoryx_hoofs/iceoryx_hoofs_types.hpp"
 
 #include <cstdio>
